#32 [chore] Move Model Files Location
resolved: #32
