--- app/src/main/java/com/mate/baedalmate/data/datasource/remote/recruit/Dormitory.kt
@@ -1,3 +1,0 @@
-package com.mate.baedalmate.data.datasource.remote.recruit
-
-enum class Dormitory { BURAM, KB, NURI, SULIM, SUNGLIM }
