--- app/src/main/java/com/mate/baedalmate/data/datasource/remote/recruit/RecruitFinishCriteria.kt
@@ -1,3 +1,0 @@
-package com.mate.baedalmate.data.datasource.remote.recruit
-
-enum class RecruitFinishCriteria { NUMBER, PRICE, TIME }
