+++ app/src/main/java/com/mate/baedalmate/data/datasource/remote/recruit/RecruitListResponse.kt
@@ -1,6 +1,7 @@
 package com.mate.baedalmate.data.datasource.remote.recruit
 
 import com.google.gson.annotations.SerializedName
+import com.mate.baedalmate.domain.model.RecruitList
 
 data class RecruitListResponse (
     @SerializedName("RecruitList")
