+++ app/src/main/java/com/mate/baedalmate/domain/model/DeliveryPlatform.kt
@@ -1,3 +1,3 @@
-package com.mate.baedalmate.data.datasource.remote.recruit
+package com.mate.baedalmate.domain.model
 
 enum class DeliveryPlatform { BAEMIN, YOGIYO, COUPANG, DDGNGYO, ETC }
