+++ app/src/main/java/com/mate/baedalmate/domain/model/Dormitory.kt
@@ -1,0 +1,3 @@
+package com.mate.baedalmate.domain.model
+
+enum class Dormitory { BURAM, KB, NURI, SULIM, SUNGLIM }
