+++ app/src/main/java/com/mate/baedalmate/domain/model/RecruitFinishCriteria.kt
@@ -1,0 +1,3 @@
+package com.mate.baedalmate.domain.model
+
+enum class RecruitFinishCriteria { NUMBER, PRICE, TIME }
