+++ app/src/main/java/com/mate/baedalmate/presentation/adapter/HomeTopPostAdapter.kt
@@ -9,7 +9,7 @@ import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.mate.baedalmate.R
-import com.mate.baedalmate.data.datasource.remote.recruit.RecruitDto
+import com.mate.baedalmate.domain.model.RecruitDto
 import com.mate.baedalmate.databinding.ItemHomeTopPostBinding
