+++ app/src/main/java/com/mate/baedalmate/presentation/viewmodel/WriteViewModel.kt
@@ -6,13 +6,14 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.mate.baedalmate.BuildConfig
 import com.mate.baedalmate.common.ListLiveData
-import com.mate.baedalmate.data.datasource.remote.recruit.DeliveryPlatform
-import com.mate.baedalmate.data.datasource.remote.recruit.Dormitory
-import com.mate.baedalmate.data.datasource.remote.recruit.MenuDto
-import com.mate.baedalmate.data.datasource.remote.recruit.PlaceDto
-import com.mate.baedalmate.data.datasource.remote.recruit.RecruitFinishCriteria
-import com.mate.baedalmate.data.datasource.remote.recruit.ShippingFeeDto
-import com.mate.baedalmate.data.datasource.remote.recruit.TagDto
+import com.mate.baedalmate.data.datasource.remote.write.CreateRecruitRequest
+import com.mate.baedalmate.domain.model.DeliveryPlatform
+import com.mate.baedalmate.domain.model.Dormitory
+import com.mate.baedalmate.domain.model.MenuDto
+import com.mate.baedalmate.domain.model.PlaceDto
+import com.mate.baedalmate.domain.model.RecruitFinishCriteria
+import com.mate.baedalmate.domain.model.ShippingFeeDto
+import com.mate.baedalmate.domain.model.TagDto
 import com.mate.baedalmate.data.datasource.remote.write.PlaceMeta
 import com.mate.baedalmate.data.datasource.remote.write.RegionInfo
 import com.mate.baedalmate.data.datasource.remote.write.ResultSearchKeyword
